refactor(api): :recycle: implement apikey #1813

+++ api/controllers/authors.js
@@ -2,6 +2,7 @@
 
 const { ObjectId } = require('mongodb')
 const COLLECTIONS = require('../constants/collections.js')
+const { validateAPIKey } = require('../services/utils')
 
 async function listAuthors(request, reply) {
   const authors = this.mongo.db.collection(COLLECTIONS.AUTHORS)
@@ -11,13 +12,21 @@ async function listAuthors(request, reply) {
 }
 
 async function addAuthor(request, reply) {
+  const { headers, body } = request
   const authors = this.mongo.db.collection(COLLECTIONS.AUTHORS)
   const id = new ObjectId()
-  const { name } = request.body
+  const { name } = body
   const data = {
     name,
     id
   }
+  const { auth } = headers
+
+  //check if the user is authorized
+  if (!validateAPIKey(auth)) {
+    return reply.code(403).send('Not authorized!')
+  }
+
   const result = await authors.insertOne(data)
 
   reply.code(201).send(result.insertedId)
@@ -35,14 +44,23 @@ async function getAuthor(request, reply) {
 }
 
 async function updateAuthor(request, reply) {
+  const { headers, body, params } = request
+
   const authors = this.mongo.db.collection(COLLECTIONS.AUTHORS)
-  const { name } = request.body
+  const { name } = body
+  const { auth } = headers
+
+  //check if the user is authorized
+  if (!validateAPIKey(auth)) {
+    return reply.code(403).send('Not authorized!')
+  }
+
   const updateDocument = {
     $set: {
       name
     }
   }
-  const result = await authors.updateOne({ id: new ObjectId(request.params.id) }, updateDocument, { upsert: true })
+  const result = await authors.updateOne({ id: new ObjectId(params.id) }, updateDocument, { upsert: true })
 
   reply.send({
     message: `${result.matchedCount} document(s) matched the filter, updated ${result.modifiedCount} document(s)`
@@ -50,6 +68,13 @@ async function updateAuthor(request, reply) {
 }
 
 async function deleteAuthor(request, reply) {
+  const { auth } = request.headers
+
+  //check if the user is authorized
+  if (!validateAPIKey(auth)) {
+    return reply.code(403).send('Not authorized!')
+  }
+
   const authors = this.mongo.db.collection(COLLECTIONS.AUTHORS)
   const result = await authors.deleteOne({ id: new ObjectId(request.params.id) })
